Cleaned up codebase
