+++ .replit
@@ -1,2 +1,2 @@
 language = "python3"
-run = "python hockey_bot.py"
+run = "bash run.sh"
