+++ README.md
@@ -1,2 +1,12 @@
 # HockeyBot
 Discord bot for hockey stats
+
+To Use:
+
+Add bot to a discord server
+Send a message in the format stat <Team> <Player>
+Example:
+  stat Leafs Marner
+
+
+
