+++ get_team.py
@@ -12,3 +12,13 @@ def get_team_id(team_name):
   obj = json.loads(data)
   id = obj[team_name]
   return id
+
+def get_team_stats(team_num):
+  url = "https://statsapi.web.nhl.com/api/v1/teams/" + str(team_num)
+
+  response = requests.get(url)
+  if response.status_code != 200:
+      print("could not retrieve team")
+      exit
+    
+  return
