+++ run.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+apt-get update
+pip install discord.py
+pip install -U python-dotenv 
+
+python hockey_bot.py
